insights reorganization
