+++ captum/insights/attr_vis/__init__.py
@@ -1,0 +1,1 @@
+from .app import AttributionVisualizer, Batch  # noqa
