+++ captum/insights/attr_vis/frontend/src/components/LabelButton.tsx
