+++ captum/insights/attr_vis/frontend/src/models/insightsConfig.ts
