+++ captum/insights/attr_vis/frontend/src/models/visualizationOutput.ts
