+++ captum/insights/attr_vis/frontend/src/utils/color.ts
