+++ captum/insights/attr_vis/widget/widget.py
@@ -3,7 +3,7 @@ import ipywidgets as widgets
 from traitlets import Dict, Instance, List, Unicode, observe
 
 from captum.insights import AttributionVisualizer
-from captum.insights.server import namedtuple_to_dict
+from captum.insights.attr_vis.server import namedtuple_to_dict
 
 
 @widgets.register
