+++ tests/insights/test_contribution.py
@@ -6,8 +6,9 @@ from typing import Callable, List, Union
 import torch
 import torch.nn as nn
 
-from captum.insights import AttributionVisualizer, Batch, FilterConfig
-from captum.insights.features import BaseFeature, FeatureOutput, ImageFeature
+from captum.insights import AttributionVisualizer, Batch
+from captum.insights.attr_vis.app import FilterConfig
+from captum.insights.attr_vis.features import BaseFeature, FeatureOutput, ImageFeature
 from tests.helpers.basic import BaseTest
